dataset/Boreas/utils.py
import cv2
import torch
import numpy as np
from torchvision.transforms import transforms
from PIL import Image


def find_nearest_ndx(ts, timestamps):
    ndx = np.searchsorted(timestamps, ts)
    if ndx == 0:
        return ndx
    elif ndx == len(timestamps):
        return ndx - 1
    else:
        assert timestamps[ndx-1] <= ts <= timestamps[ndx]
        if ts - timestamps[ndx-1] < timestamps[ndx] - ts:
            return ndx - 1
        else:
            return ndx


def load_radar_polar_pool(file_pathname, w=225, h=50):
    radar_img = cv2.imread(file_pathname, 0).transpose(1, 0)[:1343, :]
    radar_img = radar_img.astype(np.float32)
    radar_img = radar_img[np.newaxis, :, :]
    radar_img = torch.from_numpy(radar_img)
    radar_img = radar_img.unsqueeze(0)

    img_transforms = transforms.Compose([transforms.Resize((h, w))])

    downsampled_radar_img = img_transforms(radar_img)

    downsampled_radar_img = downsampled_radar_img.squeeze(0).squeeze(0).numpy().astype(np.uint8)

    return downsampled_radar_img


def load_pc(file_pathname):
    pc = np.fromfile(file_pathname, dtype=np.float32)
    pc = np.reshape(pc, (-1, 6))[:, :3]

    dist_mask = (np.linalg.norm(pc[:, :3], 2, axis=1) > 3.0) & (np.linalg.norm(pc[:, :3], 2, axis=1) < 80.0)
    pc = pc[dist_mask]

    mask = pc[:, 2] > -1.5
    pc = pc[mask]

    # coord alignment
    theta_degrees = 133.026723
    theta_radians = np.radians(theta_degrees)
    transformation_matrix = np.array([[np.cos(theta_radians), -np.sin(theta_radians), 0., 0.],
                                      [np.sin(theta_radians), np.cos(theta_radians), 0, 0.],
                                      [0., 0., 1., 0.],
                                      [0., 0., 0., 1.]])
    pc = transformation_matrix.dot(np.concatenate([pc, np.ones_like(pc[:, 0:1])], axis=-1).T).T[:, :3]
    return pc


def filter_by_yaw(pos_index_array, query_index_n_pos, database_index_n_pos, angle_threshold):
    angle_threshold = angle_threshold * np.pi / 180
    filtered_pos_index_array = []

    for i in range(len(pos_index_array)):
        query_yaw = query_index_n_pos[i, 3]
        candidate_indices = pos_index_array[i]

        if len(candidate_indices) == 0:
            filtered_pos_index_array.append(np.array([], dtype=int))
            continue

        candidate_yaws = database_index_n_pos[candidate_indices, 3]

        angle_diffs = np.abs(query_yaw - candidate_yaws)
        min_angles = np.minimum(angle_diffs, 2 * np.pi - angle_diffs)

        valid_mask = min_angles < angle_threshold
        filtered_indices = candidate_indices[valid_mask]

        filtered_pos_index_array.append(filtered_indices)
    return np.array(filtered_pos_index_array, dtype=object)
